refactor(launcher): Split eamuse stuff into module
Just moves everything eamuse related from main to a separate
module as a first step. Doesn't address ownership of options/config
or other stuff such as parameter/config overriding, yet.
